Redirect /view to /mock (for now?)
This is a trial to see how this affects conversion. Mock is
the new version, so it's more correct, but it's also nice
to have the old versions available, and nice for old links
to continue to show the same content. Redirecting for now
anyway, might just become a 'Check out HTTP Mock' banner
instead later on, TBC.