Downgrade styled-components to fix SPA link bug
Without this, if you load a blog post and then visit the home page (on
an SSR build, not in dev) then the styles are messed up, e.g. the 3 'one
click' points don't appear. Seems like something gets confused and puts
the wrong styles on the wrong elements, it's pretty bad. Sticking with
SC 4 fixes it, at no serious cost for us.

Original upgrade was as part of WillItCORS. No clear need for it though,
and all WiC functionality still seems to work fine with this.