Quick fix for background gradient overflows
The background gradients trigger horizontal page scrolling on
medium-sized screens. Only triggered by the left hand side gradients,
but just because overflowing left clips automatically.

If this causes any issues, it should be probably possible to fix with an
intermediate div to capture and hide out-of-container overflow, but it's
just a bit messy and repetitive, and I'm _reasonably_ hopeful that we
don't have any intentional whole-page horizontal scrolling where this'll
cause issues.