Update posthog and move to anonymized events
We don't identify users anyway, and this way we take that further to not
even store the anonymous person profiles at all. Posthog also helpfully
charges much less for events without profiles attached. The one identify
case that did exist was for the web post-purchase thank you page. This
attempted to link to the checkoutId from before the purchase to measure
conversion but it actually hasn't worked for a long time (since PayPro
was set up I think?) since we never use the checkoutId beforehand.

For now, this is disabled. In future if we want to track checkout
conversion based on website parameters (e.g. by referrer) then we might
need to find a new solution for that, but happy to ignore it for now.